requirements.txt
azure-storage-blob==12.8.1
Flask==2.0.1
Werkzeug
flask-bootstrap
flask-nav
flask-debug